Update Strike Finance Adapter (#14595)
